+++ primefaces/src/main/java/org/primefaces/component/datatable/export/DataTableXMLExporter.java
@@ -38,6 +38,7 @@ import org.primefaces.component.datatable.DataTable;
 import org.primefaces.component.export.ExportConfiguration;
 import org.primefaces.util.ComponentUtils;
 import org.primefaces.util.EscapeUtils;
+import org.primefaces.util.LangUtils;
 
 public class DataTableXMLExporter extends DataTableExporter {
 
@@ -90,12 +91,14 @@ public class DataTableXMLExporter extends DataTableExporter {
 
     @Override
     protected void preRowExport(DataTable table, Object document) {
-        ((PrintWriter) document).append("\t<" + table.getVar() + ">\n");
+        String rowtag = table.getExportRowTag() != null ? table.getExportRowTag() : table.getVar();
+        ((PrintWriter) document).append("\t<" + rowtag + ">\n");
     }
 
     @Override
     protected void postRowExport(DataTable table, Object document) {
-        ((PrintWriter) document).append("\t</" + table.getVar() + ">\n");
+        String rowtag = table.getExportRowTag() != null ? table.getExportRowTag() : table.getVar();
+        ((PrintWriter) document).append("\t</" + rowtag + ">\n");
     }
 
     @Override
@@ -117,7 +120,10 @@ public class DataTableXMLExporter extends DataTableExporter {
     }
 
     protected String getColumnTag(UIColumn column) {
-        String headerText = (column.getExportHeaderValue() != null) ? column.getExportHeaderValue() : column.getHeaderText();
+        String headerText = column.getExportTag();
+        if (LangUtils.isBlank(headerText)) {
+            headerText = (column.getExportHeaderValue() != null) ? column.getExportHeaderValue() : column.getHeaderText();
+        }
         UIComponent facet = column.getFacet("header");
         String columnTag;